+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 *Major*
 
 * Renamed `turn` to `stage`. ([#70](https://github.com/hjwylde/werewolf/issues/70))
+* Renamed `VillagersTurn` to `VillagesTurn`. ([#70](https://github.com/hjwylde/werewolf/issues/70))
 
 #### v0.3.4.0